chore: update Utopia package dependency and rename validation script
- Changed the Utopia package dependency in package-lock.json and package.json to reference the local tarball for version 2.10.0.
- Renamed the validation script from validate-design-system.cjs to design-system-validator.cjs.
- Removed the outdated validate-design-system.cjs file.
