+++ apps/utopia-website/package.json
@@ -19,11 +19,11 @@
     "validate": "run-p type-check lint:strict",
     "validate:fix": "run-s lint type-check",
     "check-tokens": "echo '🎨 Vérification des tokens du design system...' && npm run lint:check",
-    "validate:design-system": "node validate-design-system.cjs",
+    "validate:design-system": "node design-system-validator.cjs",
     "validate:full": "npm run validate:design-system"
   },
   "dependencies": {
-    "@club-employes/utopia": "^2.0.1",
+    "@club-employes/utopia": "file:../../packages/utopia/club-employes-utopia-2.10.0.tgz",
     "pinia": "^3.0.3",
     "vue": "^3.5.18",
     "vue-router": "^4.5.1"
