--- apps/utopia-website/validate-design-system.cjs
@@ -1,212 +1,0 @@
-#!/usr/bin/env node
-
-/**
- * Script de validation du design system Utopia
- * Vérifie que tous les tokens sont utilisés correctement
- */
-
-const fs = require('fs')
-const path = require('path')
-const { execSync } = require('child_process')
-
-console.log('🎨 Validation du Design System Utopia\n')
-
-// Couleurs pour la console
-const colors = {
-  green: '\x1b[32m',
-  red: '\x1b[31m',
-  yellow: '\x1b[33m',
-  blue: '\x1b[34m',
-  reset: '\x1b[0m',
-  bold: '\x1b[1m'
-}
-
-const log = {
-  success: (msg) => console.log(`${colors.green}✅ ${msg}${colors.reset}`),
-  error: (msg) => console.log(`${colors.red}❌ ${msg}${colors.reset}`),
-  warning: (msg) => console.log(`${colors.yellow}⚠️  ${msg}${colors.reset}`),
-  info: (msg) => console.log(`${colors.blue}ℹ️  ${msg}${colors.reset}`),
-  title: (msg) => console.log(`${colors.bold}${colors.blue}🔍 ${msg}${colors.reset}\n`)
-}
-
-// Fonction pour exécuter des commandes
-function runCommand(command, description) {
-  try {
-    log.info(`${description}...`)
-    execSync(command, { stdio: 'pipe' })
-    log.success(`${description} - OK`)
-    return true
-  } catch (error) {
-    log.error(`${description} - ÉCHEC`)
-    console.log(error.stdout?.toString())
-    console.log(error.stderr?.toString())
-    return false
-  }
-}
-
-// Analyser les fichiers CSS pour détecter les valeurs en dur
-function analyzeHardcodedValues() {
-  log.title('Analyse des valeurs CSS en dur')
-  
-  const srcDir = path.join(__dirname, 'src')
-  const hardcodedPatterns = [
-    /\d+px(?!\))/g,
-    /\d+rem(?!\))/g,
-    /\d+em(?!\))/g,
-    /#[0-9a-fA-F]{3,8}(?![a-fA-F0-9])/g,
-    /rgba?\(/g,
-    /hsla?\(/g
-  ]
-  
-  let issues = []
-  
-  function scanDirectory(dir) {
-    const files = fs.readdirSync(dir)
-    
-    for (const file of files) {
-      const filePath = path.join(dir, file)
-      const stat = fs.statSync(filePath)
-      
-      if (stat.isDirectory() && !file.startsWith('.') && file !== 'node_modules') {
-        scanDirectory(filePath)
-      } else if (file.endsWith('.vue') || file.endsWith('.css') || file.endsWith('.scss')) {
-        const content = fs.readFileSync(filePath, 'utf8')
-        
-        hardcodedPatterns.forEach(pattern => {
-          const matches = content.match(pattern)
-          if (matches) {
-            issues.push({
-              file: filePath.replace(__dirname + '/', ''),
-              matches: matches,
-              pattern: pattern.source
-            })
-          }
-        })
-      }
-    }
-  }
-  
-  scanDirectory(srcDir)
-  
-  if (issues.length === 0) {
-    log.success('Aucune valeur CSS en dur détectée')
-  } else {
-    log.warning(`${issues.length} fichier(s) avec des valeurs en dur détectées:`)
-    issues.forEach(issue => {
-      console.log(`  📄 ${issue.file}:`)
-      issue.matches.forEach(match => {
-        console.log(`    - ${match}`)
-      })
-    })
-  }
-  
-  return issues.length === 0
-}
-
-// Vérifier l'utilisation des tokens Utopia
-function checkTokenUsage() {
-  log.title('Vérification de l\'utilisation des tokens')
-  
-  // Lire les tokens disponibles
-  const tokensFile = path.join(__dirname, '../../../packages/utopia/src/tokens/generated/club-employes/light.css')
-  
-  if (!fs.existsSync(tokensFile)) {
-    log.warning('Fichier de tokens non trouvé, vérification ignorée')
-    return true
-  }
-  
-  const tokensContent = fs.readFileSync(tokensFile, 'utf8')
-  const tokenMatches = tokensContent.match(/--[\w-]+:/g)
-  const availableTokens = tokenMatches ? tokenMatches.map(t => t.replace(':', '')) : []
-  
-  log.info(`${availableTokens.length} tokens disponibles dans le design system`)
-  
-  // Compter l'utilisation des tokens dans le code
-  const srcDir = path.join(__dirname, 'src')
-  let tokenUsage = {}
-  
-  function countTokens(dir) {
-    const files = fs.readdirSync(dir)
-    
-    for (const file of files) {
-      const filePath = path.join(dir, file)
-      const stat = fs.statSync(filePath)
-      
-      if (stat.isDirectory() && !file.startsWith('.')) {
-        countTokens(filePath)
-      } else if (file.endsWith('.vue') || file.endsWith('.css')) {
-        const content = fs.readFileSync(filePath, 'utf8')
-        
-        availableTokens.forEach(token => {
-          const regex = new RegExp(`var\\(${token}\\)`, 'g')
-          const matches = content.match(regex)
-          if (matches) {
-            tokenUsage[token] = (tokenUsage[token] || 0) + matches.length
-          }
-        })
-      }
-    }
-  }
-  
-  countTokens(srcDir)
-  
-  const usedTokens = Object.keys(tokenUsage)
-  const usagePercentage = (usedTokens.length / availableTokens.length * 100).toFixed(1)
-  
-  log.info(`${usedTokens.length}/${availableTokens.length} tokens utilisés (${usagePercentage}%)`)
-  
-  if (usedTokens.length > 0) {
-    log.success('Tokens du design system utilisés correctement')
-  } else {
-    log.warning('Aucun token du design system détecté')
-  }
-  
-  return true
-}
-
-// Exécution principale
-async function main() {
-  let allPassed = true
-  
-  // 1. Type checking TypeScript
-  log.title('Vérification TypeScript')
-  if (!runCommand('npm run type-check', 'Type checking')) {
-    allPassed = false
-  }
-  
-  // 2. Linting ESLint
-  log.title('Vérification ESLint')
-  if (!runCommand('npm run lint:check', 'Linting')) {
-    allPassed = false
-  }
-  
-  // 3. Analyse des valeurs en dur
-  if (!analyzeHardcodedValues()) {
-    allPassed = false
-  }
-  
-  // 4. Vérification des tokens
-  if (!checkTokenUsage()) {
-    allPassed = false
-  }
-  
-  // 5. Build test
-  log.title('Test de build')
-  if (!runCommand('npm run build', 'Build de production')) {
-    allPassed = false
-  }
-  
-  console.log('\n' + '='.repeat(60))
-  
-  if (allPassed) {
-    log.success('🎉 TOUTES LES VALIDATIONS SONT PASSÉES !')
-    log.success('Le code respecte parfaitement le design system Utopia')
-    process.exit(0)
-  } else {
-    log.error('❌ ÉCHEC DE LA VALIDATION')
-    log.error('Corrigez les erreurs ci-dessus avant de continuer')
-    process.exit(1)
-  }
-}
-
-main().catch(console.error)
